Consolidates scripts into single module that supports CLI and programmatic invocation
